SOURCE/AZ-EVENT-HUB-TO-KAFKA/requirements.txt
termcolor
confluent-kafka
azure-eventhub